requirements.txt
flask
flask_sqlalchemy
flask-restful
flask-marshmallow
marshmallow_sqlalchemy
flask_script
flask_migrate
pymysql
faker
bcrypt
pyjwt
pytz
python-dateutil